+++ NineZeroProject/UI/Profile/HTProfileRootController.m
@@ -189,7 +189,7 @@
 - (UICollectionViewCell *)collectionView:(UICollectionView *)collectionView cellForItemAtIndexPath:(NSIndexPath *)indexPath {
     HTProfileRecordCell *cell = [collectionView dequeueReusableCellWithReuseIdentifier:NSStringFromClass([HTProfileRecordCell class]) forIndexPath:indexPath];
     if (_profileInfo.answer_list.count != 0) {
-        DLog(@"Index:%ld", indexPath.row)
+        DLog(@"Index:%ld", (long)indexPath.row)
         DLog(@"answerList->%@", _profileInfo.answer_list[indexPath.row].question_video_cover);
         [cell.coverImageView sd_setImageWithURL:[NSURL URLWithString:_profileInfo.answer_list[indexPath.row].question_video_cover] placeholderImage:[UIImage imageNamed:@"img_monday_music_cover_default"]];
     }
